refactored Settings Provider and added a simple Test for it
